requirements.txt
Pillow>=9.0.0
exifread>=3.0.0
mutagen>=1.47.0
PyYAML>=6.0